Added agent placer
Added agent placer. I have commented out few lines in
EntryPointImpl.setAgentPositions() to bypass write transaction check.
This causes the string value of Agent Positions in properties view is
not updated in real time. User have to click out and click back to the
agent to see the position changes. Also, currently you can't unplace an
agent xd. It has to be a small bug